die will align to assigned face value
A die will instantly align to the value it is given by the roll command. I couldn't figure out how to get it to correctly transition by increments.
Also the RNG is super broken, its giving the same set of numbers every time.

+++ game.h
@@ -27,8 +27,8 @@ struct yahtzee_game
 	// END State Settings
 
 	//START Physics Settings
-	int throw_variance_floor;
-	int throw_variance_ceiling;
+	double throw_variance_floor;
+	double throw_variance_ceiling;
 	float throw_entropy;
 	float throw_force;
 	float gravity_force;
